refactor: move sidebar layout support
