+++ Sources/Sidebar/SidebarLayoutSupport.swift
@@ -1,0 +1,168 @@
+import AppKit
+import SwiftUI
+
+struct SidebarResizerAccessibilityModifier: ViewModifier {
+    let accessibilityIdentifier: String?
+
+    @ViewBuilder
+    func body(content: Content) -> some View {
+        if let accessibilityIdentifier {
+            content.accessibilityIdentifier(accessibilityIdentifier)
+        } else {
+            content
+        }
+    }
+}
+
+struct SidebarTopScrim: View {
+    let height: CGFloat
+
+    var body: some View {
+        SidebarTopBlurEffect()
+            .frame(height: height)
+            .mask(
+                LinearGradient(
+                    colors: [
+                        Color.black.opacity(0.95),
+                        Color.black.opacity(0.75),
+                        Color.black.opacity(0.35),
+                        Color.clear
+                    ],
+                    startPoint: .top,
+                    endPoint: .bottom
+                )
+            )
+    }
+}
+
+private struct SidebarTopBlurEffect: NSViewRepresentable {
+    func makeNSView(context: Context) -> NSVisualEffectView {
+        let view = NSVisualEffectView()
+        view.blendingMode = .withinWindow
+        view.material = .underWindowBackground
+        view.state = .active
+        view.isEmphasized = false
+        return view
+    }
+
+    func updateNSView(_ nsView: NSVisualEffectView, context: Context) {}
+}
+
+struct SidebarScrollViewResolver: NSViewRepresentable {
+    let onResolve: (NSScrollView?) -> Void
+
+    func makeNSView(context: Context) -> SidebarScrollViewResolverView {
+        let view = SidebarScrollViewResolverView()
+        view.onResolve = onResolve
+        return view
+    }
+
+    func updateNSView(_ nsView: SidebarScrollViewResolverView, context: Context) {
+        nsView.onResolve = onResolve
+        nsView.resolveScrollView()
+    }
+}
+
+final class SidebarScrollViewResolverView: NSView {
+    var onResolve: ((NSScrollView?) -> Void)?
+
+    override func viewDidMoveToSuperview() {
+        super.viewDidMoveToSuperview()
+        resolveScrollView()
+    }
+
+    override func viewDidMoveToWindow() {
+        super.viewDidMoveToWindow()
+        resolveScrollView()
+    }
+
+    func resolveScrollView() {
+        DispatchQueue.main.async { [weak self] in
+            guard let self else { return }
+            onResolve?(self.enclosingScrollView)
+        }
+    }
+}
+
+enum SidebarPathFormatter {
+    static let homeDirectoryPath: String = FileManager.default.homeDirectoryForCurrentUser.path
+
+    static func shortenedPath(
+        _ path: String,
+        homeDirectoryPath: String = Self.homeDirectoryPath
+    ) -> String {
+        let trimmed = path.trimmingCharacters(in: .whitespacesAndNewlines)
+        guard !trimmed.isEmpty else { return path }
+        if trimmed == homeDirectoryPath {
+            return "~"
+        }
+        if trimmed.hasPrefix(homeDirectoryPath + "/") {
+            return "~" + trimmed.dropFirst(homeDirectoryPath.count)
+        }
+        return trimmed
+    }
+}
+
+enum SidebarTrailingAccessoryWidthPolicy {
+    static let closeButtonWidth: CGFloat = 16
+
+    static func width(
+        canCloseWorkspace: Bool
+    ) -> CGFloat {
+        return canCloseWorkspace ? closeButtonWidth : 0
+    }
+}
+
+struct ClearScrollBackground: ViewModifier {
+    func body(content: Content) -> some View {
+        if #available(macOS 13.0, *) {
+            content
+                .scrollContentBackground(.hidden)
+                .background(ScrollBackgroundClearer())
+        } else {
+            content
+                .background(ScrollBackgroundClearer())
+        }
+    }
+}
+
+private struct ScrollBackgroundClearer: NSViewRepresentable {
+    func makeNSView(context: Context) -> NSView {
+        NSView()
+    }
+
+    func updateNSView(_ nsView: NSView, context: Context) {
+        DispatchQueue.main.async {
+            guard let scrollView = findScrollView(startingAt: nsView) else { return }
+            // Clear all backgrounds and mark as non-opaque for transparency.
+            scrollView.drawsBackground = false
+            scrollView.backgroundColor = .clear
+            scrollView.wantsLayer = true
+            scrollView.layer?.backgroundColor = NSColor.clear.cgColor
+            scrollView.layer?.isOpaque = false
+
+            scrollView.contentView.drawsBackground = false
+            scrollView.contentView.backgroundColor = .clear
+            scrollView.contentView.wantsLayer = true
+            scrollView.contentView.layer?.backgroundColor = NSColor.clear.cgColor
+            scrollView.contentView.layer?.isOpaque = false
+
+            if let docView = scrollView.documentView {
+                docView.wantsLayer = true
+                docView.layer?.backgroundColor = NSColor.clear.cgColor
+                docView.layer?.isOpaque = false
+            }
+        }
+    }
+
+    private func findScrollView(startingAt view: NSView) -> NSScrollView? {
+        var current: NSView? = view
+        while let candidate = current {
+            if let scrollView = candidate as? NSScrollView {
+                return scrollView
+            }
+            current = candidate.superview
+        }
+        return nil
+    }
+}
